fixed bug binom distr

+++ main.c
@@ -51,7 +51,7 @@ int main(int argc, char* const* argv){
 				break ; 
 			
 			case 'r':
-				p2 = atoi(optarg) ; 
+				p2 = atof(optarg) ; 
 				break ;
 
 			case 'n':
